+++ internal/daemon/docker_monitor/client/bridge.go
@@ -1,11 +1,12 @@
 package client
 
 import (
+	"encoding/json"
 	"fmt"
 	"strings"
 )
 
-func (d *docker) Bridges() ([]string, error) {
+func (d *docker) bridges() ([]string, error) {
 	args := []string{"network", "ls", "-q", "--filter", "Driver=bridge"}
 	result, err := d.command(args...)
 	if err != nil {
@@ -25,42 +26,17 @@ func (d *docker) Bridges() ([]string, error) {
 	return lines, nil
 }
 
-func (d *docker) BridgeNames() ([]string, error) {
-	bridges, err := d.Bridges()
-	if err != nil {
-		return nil, err
-	}
-
-	var names []string
-
-	for _, bridge := range bridges {
-		bridgeName, err := d.BridgeName(bridge)
-		if err != nil {
-			d.logger.Error(err.Error())
-			continue
-		}
-		names = append(names, bridgeName)
-	}
-
-	return names, nil
-}
-
-func (d *docker) BridgeName(bridgeID string) (string, error) {
-	format := fmt.Sprintf(`{{"br-%s" | or (index .Options "com.docker.network.bridge.name")}}`, bridgeID)
-	args := []string{"network", "inspect", "-f", format, bridgeID}
+func (d *docker) bridgeInfo(bridgeID string) (DockerBridgeInspect, error) {
+	args := []string{"network", "inspect", bridgeID}
 	result, err := d.command(args...)
 	if err != nil {
-		return "", fmt.Errorf("failed to get bridge name: %s", err.Error())
+		return DockerBridgeInspect{}, fmt.Errorf("failed to get bridge name: %s", err.Error())
 	}
-	return strings.TrimSpace(string(result)), nil
-}
 
-func (d *docker) BridgeSubnet(bridgeID string) (string, error) {
-	format := fmt.Sprintf(`{{range .IPAM.Config}}{{.Subnet}}{{end}}`)
-	args := []string{"network", "inspect", "-f", format, bridgeID}
-	result, err := d.command(args...)
-	if err != nil {
-		return "", fmt.Errorf("failed to get bridge subnet: %s", err.Error())
+	var info []DockerBridgeInspect
+	if err := json.Unmarshal(result, &info); err != nil {
+		return DockerBridgeInspect{}, err
 	}
-	return strings.TrimSpace(string(result)), nil
+
+	return info[0], nil
 }