+++ internal/daemon/docker_monitor/config.go
@@ -1,0 +1,13 @@
+package docker_monitor
+
+type Config struct {
+	Path         string
+	RuleStrategy RuleStrategy
+}
+
+type RuleStrategy int8
+
+const (
+	RuleStrategyRebuild RuleStrategy = iota + 1
+	RuleStrategyIncremental
+)